Added a new interface to SwsContext that allows you to scale directly to a java array, avoiding 1 redundant copy. Updated JJMediaReader to use this new interface when getting a BufferedImage. Updated JJFileChooser to use JJMediaReader. Fixed a bug in AVPlane where it didn't set the byte endianness properly. Hid some of the debug output.

+++ trunk/jjmpeg/src/au/notzed/jjmpeg/AVRational.java
@@ -46,7 +46,7 @@ public class AVRational extends AVRationalAbstract {
 	public static final AVRational AV_TIME_BASE_Q = create(1, 1000000);
 
 	public double q2d() {
-		return getNum() / (double) getDen();
+		return (double) getNum() / (double) getDen();
 	}
 
 	static public long rescaleQ(long a, AVRational bq, AVRational cq) {
